requirements.txt
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
cryptography==40.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
pycparser==2.21
python-telegram-bot==20.3
sniffio==1.3.0
